chore(urls): keep root routing declarative
